requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate